+++ docs_src/src/guide/changelog.md
@@ -6,6 +6,7 @@
 
 - Added `Request.Request()` accessor to get the raw `*http.Request`.
 - Fixed a bug allowing non-core middleware applied to the root router to be executed when the "Not Found" or "Method Not Allowed" routes were matched.
+- Fixed a bug making route groups (subrouters with empty prefix) conflict with their parent router when two routes having the same path but different methods are registered in both routers.
 - Added [chained writers](./basics/responses.html#chained-writers).
 - Added [gzip compression middleware](./basics/middleware.html#gzip).
 - Added the ability to register route-specific middleware in `Router.Static()`.